// src/common/enums/error-code.enum.ts
export enum ErrorCode {
  // 通用错误 1xxx
  UNKNOWN_ERROR = 1000,
  INVALID_PARAMS = 1001,
  UNAUTHORIZED = 1002,
  FORBIDDEN = 1003,
  NOT_FOUND = 1004,
  METHOD_NOT_ALLOWED = 1005,
  REQUEST_TIMEOUT = 1006,
  TOO_MANY_REQUESTS = 1007,

  // 业务错误 2xxx
  BUSINESS_ERROR = 2000,

  // 数据库错误 3xxx
  DATABASE_ERROR = 3000,
  DUPLICATE_ENTRY = 3001,
}